refactor(permission): 提取权限主体纯逻辑
